socket: make dispatch infallible, except for emit errors.
socket dispatch can't be allowed to fail, it's too late to handle
errors there. For example, you can push a malformed packet to a
raw::Socket. Its dispatch() would detect that and error out, but *not*
pop the packet from the queue, resulting in the socket getting stuck forever!

Instead, all errors are now logged (and the offending packet is dropped if applicable).

Also, Err(Exhausted) was returned as "no error, but I have no packets to emit", but
that's not really useful to Interface, because it can already see if the
socket emitted packets or not.