Use renewal time from DHCP server ACK, if given
Per RFC 2132 section 9.11 the server can manually specify a renewal (T1)
time different from the default of half the lease time through option
code 58. This PR updates the behavior of the dhcp client to use that
value, if provided, and only if not provided does it default to half of
the lease duration.

Since the current state of smoltcp does not seem to follow the REBINDING
state, I also made it look for a value in option code 59 (which should
be the rebinding (T2) interval) and use that if no T1 interval interval
is provided. This behavior seems sensible to me, given that we're not
following the REBINDING part of the spec, but I can change it to ignore
option code 59, or any other handling, if that is preferred.